# EXPERIMENTS: shorten cache lookup chains, invalidate cache on new ensures