--- a/booster/library/Booster/Pattern/Rewrite.hs
+++ b/booster/library/Booster/Pattern/Rewrite.hs
@@ -51,7 +51,8 @@ import Booster.LLVM as LLVM (API)
 import Booster.Log
 import Booster.Pattern.ApplyEquations (
     EquationFailure (..),
-    SimplifierCache,
+    SimplifierCache (..),
+    CacheTag (Equations),
     evaluatePattern,
     simplifyConstraint,
  )
@@ -400,6 +401,13 @@ applyRule pat@Pattern{ceilConditions} rule =
                             pure ()
                         Left other ->
                             liftIO $ Exception.throw other
+
+                    -- if a new constraint is going to be added, the equation cache is invalid
+                    unless (null newConstraints) $ do
+                        withContextFor Equations . logMessage $
+                            ("New path condition ensured, invalidating cache" :: Text)
+                        lift . RewriteT . lift . modify $ \s -> s{equations = mempty}
+
 
                     -- existential variables may be present in rule.rhs and rule.ensures,
                     -- need to strip prefixes and freshen their names with respect to variables already