@jberthold 2024-07-26
Performance for KEVM tests on par with master.

| Test                                                                  | 3982-shorten-simplification-cache-lookup-chains time | master-f0b3596f9 time | (3982-shorten-simplification-cache-lookup-chains/master-f0b3596f9) time 
|-----------------------------------------------------------------------|------------------------------------------------------|-----------------------|-------------------------------------------------------------------------
| kontrol/test-arithmetictest-test_max1_broken-uint256-uint256-0-spec.k | 38.76                                                | 40.32                 | 0.9613095238095237                                                      
| mcd/flipper-addu48u48-fail-rough-spec.k                               | 43.28                                                | 44.86                 | 0.9647793134195275                                                      
| benchmarks/ecrecoverloop02-sig0-invalid-spec.k                        | 108.67                                               | 112.53                | 0.9656980360792677                                                      
| TOTAL                                                                 | 190.71                                   | 197.71                | 0.9645946082646298