@jberthold 2024-08-02
Insignificant changes to `kevm` and `kontrol` tests

| Test                                                       | 3982-shorten-simplification-cache-lookup-chains time | master-e4bd3aadc time | (3982-shorten-simplification-cache-lookup-chains/master-e4bd3aadc) time 
|------------------------------------------------------------|------------------------------------------------------|-----------------------|-------------------------------------------------------------------------
| erc20/ds/transferFrom-failure-1-c-spec.k                   | 83.36                                                | 79.95                 | 1.0426516572858036                                                      
| kontrol/test-expectcalltest-testexpectstaticcall-0-spec.k  | 41.07                                                | 39.03                 | 1.0522674865488086                                                      
| kontrol/test-expectcalltest-testexpectregularcall-0-spec.k | 42.49                                                | 39.99                 | 1.0625156289072268                                                      
| benchmarks/encodepacked-keccak01-spec.k                    | 41.0                                                 | 38.24                 | 1.0721757322175731                                                      
| TOTAL                                                      | 207.92                                   | 197.21                | 1.054307590892957                                                       

| Test                                                                                     | 3982-shorten-simplification-cache-lookup-chains time | master-e4bd3aadc time | (3982-shorten-simplification-cache-lookup-chains/master-e4bd3aadc) time 
|------------------------------------------------------------------------------------------|------------------------------------------------------|-----------------------|-------------------------------------------------------------------------
| BlockParamsTest.testFee(uint256)                                                         | 15.44                                                | 16.98                 | 0.9093050647820965                                                      
| CounterTest.testIncrement()                                                              | 19.53                                                | 21.0                  | 0.93                                                                    
| BlockParamsTest.testChainId(uint256)                                                     | 15.43                                                | 16.57                 | 0.9312009656004828                                                      
| BytesTypeTest.test_bytes4(bytes4)                                                        | 15.32                                                | 16.42                 | 0.933008526187576                                                       
| kontrol/src/tests/integration/test_foundry_prove.py::test_foundry_remove_node            | 13.52                                                | 14.28                 | 0.9467787114845938                                                      
| AddrTest.test_notBuiltinAddress_concrete()                                               | 11.92                                                | 12.51                 | 0.9528377298161471                                                      
| kontrol/src/tests/integration/test_foundry_prove.py::test_foundry_xml_report             | 30.97                                                | 32.22                 | 0.9612042209807573                                                      
| StartPrankTestMsgSender.test_startprank_msgsender_setup()                                | 31.13                                                | 32.28                 | 0.9643742255266419                                                      
| MockFunctionTest.test_mock_function_all_args()                                           | 44.22                                                | 42.66                 | 1.0365682137834038                                                      
| AddrTest.test_builtInAddresses()                                                         | 12.77                                                | 12.23                 | 1.044153720359771                                                       
| kontrol/src/tests/integration/test_foundry_prove.py::test_foundry_auto_abstraction       | 42.42                                                | 40.55                 | 1.0461159062885328                                                      
| CoinBaseTest.test_coinbase_setup()                                                       | 28.12                                                | 26.2                  | 1.0732824427480916                                                      
| PlainPrankTest.test_prank_zeroAddress_true()                                             | 24.7                                                 | 22.78                 | 1.0842844600526778                                                      
| kontrol/src/tests/integration/test_foundry_prove.py::test_foundry_show_with_hex_encoding | 5.05                                                 | 4.65                  | 1.086021505376344                                                       
| BlockParamsTest.testRoll(uint256)                                                        | 16.82                                                | 15.45                 | 1.0886731391585762                                                      
| CounterTest.testSetNumber(uint256)                                                       | 23.0                                                 | 21.07                 | 1.0915994304698624                                                      
| AssumeTest.test_multi_assume(address,address)                                            | 22.38                                                | 20.43                 | 1.0954478707782673                                                      
| BlockParamsTest.testCoinBase()                                                           | 13.46                                                | 12.08                 | 1.1142384105960266                                                      
| TOTAL                                                                                    | 386.2                                                | 380.36   | 1.0153538752760545    |